requirements.txt
pillow~=10.2.0
pyautogui~=0.9.54
easyocr~=1.7.1
opencv-python~=4.9.0.80
matplotlib~=3.8.3
PyAutoGUI~=0.9.54
numpy~=1.26.4